+++ utils/render.js
@@ -1,0 +1,14 @@
+//Fonction de connexion à la base de données MongoDB
+exports.returnRender = function(viewTitle){
+  var obj =  {
+      title: viewTitle,
+      home: 'Accueil',
+      search: 'Recherche',
+      login: 'Connexion',
+      spot: 'Mes spots',
+      application: 'Une application pour les surfeurs à la recherche de nouveaux spot de surf!',
+      tchat: 'Chat en ligne'
+    };
+
+    return obj;
+}
